+++ src/FamilyHubs.Referral.Core/DistributedCache/IReferralDistributedCache.cs
@@ -1,0 +1,9 @@
+﻿using FamilyHubs.Referral.Core.Models;
+
+namespace FamilyHubs.Referral.Core.DistributedCache;
+
+public interface IReferralDistributedCache
+{
+    Task<ProfessionalReferralModel?> GetProfessionalReferralAsync();
+    Task SetProfessionalReferralAsync(ProfessionalReferralModel model);
+}
